pyproject.toml
[project]
name = "gce-english-be"
version = "0.1.0"
description = "Backend services for OCR ingestion of GCE English examination papers"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "loguru>=0.7.3",
    "pydantic-settings>=2.4.0",
    "pypdfium2>=4.30.0",
    "pillow>=10.3.0",
    "numpy>=1.26.0",
    "python-multipart>=0.0.9",
    "rapidocr",
    "onnxruntime",
    "openai>=1.51.0",
    "reportlab>=4.2.0",
    "jinja2>=3.1.4",
    "xhtml2pdf>=0.2.15",
    "weasyprint>=61.2",
    "playwright>=1.47.0",
    "requests>=2.32.3",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.3.0",
    "tavily-python>=0.3.0",
    "supabase>=2.0.0",
    "httpx>=0.27.0",
    "passlib==1.7.4",
    "bcrypt==4.3.0",
    "PyJWT>=2.8.0",
    "pydantic[email]>=2.10.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]
